+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Bump `pyxavi` to `v0.8.0` ([#3](https://github.com/XaviArnaus/masto-feed/pull/3))
 - Initial documentation ([#4](https://github.com/XaviArnaus/masto-feed/pull/4))
 - Make Log rotation as default ([#5](https://github.com/XaviArnaus/masto-feed/pull/5))
+- Avoid answering with ERROR_INVALID_ACTION if it's organic mention ([#8](https://github.com/XaviArnaus/masto-feed/pull/8))
 
 ### Fixed